mv dsp_aware_pruning location
