+++ hls4ml/optimization/dsp_aware_pruning/keras/regularizers.py
@@ -1,7 +1,7 @@
 import numpy as np
 import tensorflow as tf
 
-from hls4ml.optimization.config import SUPPORTED_STRUCTURES
+from hls4ml.optimization.dsp_aware_pruning.config import SUPPORTED_STRUCTURES
 
 
 @tf.keras.utils.register_keras_serializable(name='DenseRegularizer')
