+++ hls4ml/optimization/dsp_aware_pruning/keras/utils.py
