+++ hls4ml/optimization/dsp_aware_pruning/knapsack.py
