+++ hls4ml/optimization/dsp_aware_pruning/objectives/__init__.py
@@ -3,8 +3,8 @@ from abc import ABC, abstractmethod
 
 import numpy as np
 
-from hls4ml.optimization.attributes import OptimizationAttributes
-from hls4ml.optimization.config import SUPPORTED_STRUCTURES
+from hls4ml.optimization.dsp_aware_pruning.attributes import OptimizationAttributes
+from hls4ml.optimization.dsp_aware_pruning.config import SUPPORTED_STRUCTURES
 
 '''
 Pruning & weight sharing are formulated as an optimization problem, with the aim of minimizing some metric
