+++ hls4ml/optimization/dsp_aware_pruning/objectives/gpu_objectives.py
@@ -2,9 +2,9 @@ import logging
 
 import numpy as np
 
-from hls4ml.optimization.attributes import OptimizationAttributes
-from hls4ml.optimization.config import SUPPORTED_STRUCTURES
-from hls4ml.optimization.objectives import ObjectiveEstimator
+from hls4ml.optimization.dsp_aware_pruning.attributes import OptimizationAttributes
+from hls4ml.optimization.dsp_aware_pruning.config import SUPPORTED_STRUCTURES
+from hls4ml.optimization.dsp_aware_pruning.objectives import ObjectiveEstimator
 
 
 class GPUFLOPEstimator(ObjectiveEstimator):
