+++ test/pytest/test_optimization/test_keras/test_weight_sharing.py
@@ -4,10 +4,10 @@ from qkeras import QDense
 from tensorflow.keras.layers import Dense
 from tensorflow.keras.models import Sequential
 
-from hls4ml.optimization.attributes import get_attributes_from_keras_model
-from hls4ml.optimization.config import SUPPORTED_STRUCTURES
-from hls4ml.optimization.keras.masking import get_model_masks
-from hls4ml.optimization.objectives import ObjectiveEstimator
+from hls4ml.optimization.dsp_aware_pruning.attributes import get_attributes_from_keras_model
+from hls4ml.optimization.dsp_aware_pruning.config import SUPPORTED_STRUCTURES
+from hls4ml.optimization.dsp_aware_pruning.keras.masking import get_model_masks
+from hls4ml.optimization.dsp_aware_pruning.objectives import ObjectiveEstimator
 
 # Similar tests in test_masking.py, weight sharing instead of pruning
 sparsity = 0.33
